Added pothole detection algorithm and start button that will begin accelerometer tracking.  Going back away from the map stops the accelerometer
